requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-Dance[sqla]==7.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
Flask-Migrate==4.0.5  # For database migrations
Flask-WTF==1.2.1  # For form handling and CSRF protection
email-validator==2.1.0  # For email validation

# Database
SQLAlchemy==2.0.20
PyMySQL==1.1.0
cryptography==41.0.3  # Required for PyMySQL with SSL
alembic==1.12.1  # For database migrations
sqlalchemy-utils==0.41.1  # Useful SQLAlchemy utilities

# Environment and configuration
python-dotenv==1.0.0
environs==9.5.0  # Enhanced environment variable parsing with validation

# HTTP and API
requests==2.31.0
httpx==0.25.2  # Modern HTTP client with async support
tenacity==8.2.3  # For retrying failed operations (API calls, DB connections)

# Google API
google-auth==2.22.0
google-api-python-client==2.97.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0

# Email
secure-smtplib==0.2.1  # More secure SMTP library

# PDF processing
reportlab==4.0.4
PyPDF2==3.0.1
pdfkit==1.0.0
img2pdf==0.4.4
Pillow==10.0.0  # Required for image processing
weasyprint==60.1  # Alternative HTML to PDF converter

# Date and time
pytz==2023.3
python-dateutil==2.8.2  # For advanced date handling

# Utilities
urllib3==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pydantic==2.5.2  # For data validation
python-slugify==8.0.1  # For creating URL-friendly slugs
humanize==4.8.0  # For human-readable data (file sizes, dates)

# Monitoring and Error Handling
sentry-sdk[flask]==1.38.0  # Error tracking (optional)
logging-plus==0.0.7  # Enhanced logging capabilities

# Testing
pytest==7.4.3  # For unit testing
pytest-flask==1.3.0  # Flask-specific testing utilities
coverage==7.3.2  # For test coverage reporting
